+++ extensions/user/zaggino.brackets-git/src/Remotes.js
@@ -97,7 +97,7 @@ define(function (require) {
 
             // Disable Git-push and Git-pull if there are not remotes defined
             $gitPanel
-                .find(".git-pull, .git-push")
+                .find(".git-pull, .git-push, .git-fetch")
                 .prop("disabled", remotes.length === 0);
 
             // Add options to change remote
@@ -320,6 +320,38 @@ define(function (require) {
             });
     }
 
+    function handleFetch(silent) {
+
+        // Tell the rest of the plugin that the fetch has started
+        EventEmitter.emit(Events.FETCH_STARTED);
+
+        var q;
+
+        if (!silent) {
+
+            // If it's not a silent fetch show a progress window
+            q = ProgressDialog.show(Git.fetchAllRemotes())
+            .catch(function (err) {
+                ErrorHandler.showError(err);
+            })
+            .then(ProgressDialog.waitForClose);
+
+        } else {
+
+            // Else fetch in the background
+            q = Git.fetchAllRemotes()
+            .catch(function (err) {
+                ErrorHandler.logError(err);
+            });
+
+        }
+
+        // Tell the rest of the plugin that the fetch has completed
+        return q.finally(function () {
+            EventEmitter.emit(Events.FETCH_COMPLETE);
+        });
+    }
+
     // Event subscriptions
     EventEmitter.on(Events.GIT_ENABLED, function () {
         initVariables();
@@ -330,6 +362,7 @@ define(function (require) {
             remoteName  = $remote.data("remote-name"),
             type        = $remote.data("type");
         selectRemote(remoteName, type);
+        EventEmitter.emit(Events.REFRESH_COUNTERS);
     });
     EventEmitter.on(Events.HANDLE_REMOTE_CREATE, function () {
         handleRemoteCreation();
@@ -345,6 +378,9 @@ define(function (require) {
     EventEmitter.on(Events.HANDLE_PUSH, function () {
         var remoteName = $selectedRemote.data("remote");
         pushToRemote(remoteName);
+    });
+    EventEmitter.on(Events.HANDLE_FETCH, function () {
+        handleFetch();
     });
 
 });